refactor(languages): unify per-language config into a single root registry
Phase 0 of a staged refactor to make adding a language a one-directory change
instead of editing the ~11 scattered sites it touches today.

Collapse the byte-identical client/api language registries, the hand-written
LanguageCode union, and the three DEFAULT_LANGUAGE copies into one source of
truth under languages/:

- languages/{types,registry}.ts + languages/<code>/manifest.ts (af/de/es),
  one self-contained manifest per language (avoid-words inlined).
- src/lib/languages.ts, api/src/lib/languages.ts, src/types/language.ts and
  src/constants/languages.ts are now thin re-exports of the registry.
- LanguageCode is derived from the manifest keys (keyof typeof MANIFESTS), so
  it is never hand-maintained.
- The API ships as a self-contained api/ tree, so the shared dir is baked into
  the image with one line: Dockerfile `COPY languages ./languages` (resolves to
  /app/languages, where the API's ../../../languages import points at runtime).

Pure refactor: same data, same public exports, no behavior change. Verified
client tsc clean, next build compiles, 236 client + 170 API tests pass, lint 0.
