+++ languages/af/manifest.ts
@@ -1,0 +1,19 @@
+const AVOID_WORDS = new Set([
+  "'n", "die", "en", "of", "in", "op", "vir", "met", "na", "van",
+  "is", "het", "om", "te", "dat", "wat", "as", "aan", "by", "sy", "hy",
+  "nie", "ek", "jy", "ons", "hulle", "dit", "was", "sal", "kan", "moet",
+  "maar", "ook", "al", "nog", "so", "toe", "nou", "net", "eers", "dan",
+]);
+
+export const af = {
+  name: 'Afrikaans',
+  native: 'Afrikaans',
+  code: 'af' as const,
+  flag: '\u{1F1FF}\u{1F1E6}',
+  ttsCode: 'af-ZA',
+  ttsVoice: 'af-ZA-Standard-A',
+  tatoebaCode: 'afr',
+  fallbackTts: ['af', 'nl-NL', 'nl'],
+  avoidWords: AVOID_WORDS,
+  testPhrase: 'Hallo, hoe gaan dit met jou?',
+};
