+++ src/constants/languages.ts
@@ -1,3 +1,2 @@
-import { LanguageCode } from "@/types/language";
-
-export const DEFAULT_LANGUAGE: LanguageCode = 'af';
+// Re-export from the shared language registry (`languages/registry.ts`).
+export { DEFAULT_LANGUAGE } from '../../languages/registry';
